Evaluate syllable matcher (#2)
* Add MEX function for evaluating syllable

Provide a template syllable, and one or more syllables to evaluate. It returns the final score and length associated with each syllable to evaluate. Useful for generating ROC curve.

* Squash many warnings and bug

The callback pointers were not being initialized when called via MATLAB. Lead to seg faults on some calls (not all). In the process, cleaned up some warnings and other issues.

* Evaluate syllable matching via ROC curve, confusion matrix and accuracy calculation