Build templates based on many warped renditions
New MATLAB function `build_template`, that depends on private functions from the find-audio library, to warp together multiple spectrograms and create a template that can be used for matching. Also returns a set of weights that can be used to discount frequenty areas of variability (not currently used).

Updated DTM and match syllables to support loading spectral template directly.